Fixed a double to int cast

+++ moea/src/main/java/moea/Chromosome.java
@@ -15,7 +15,7 @@ public class Chromosome {
     double edgeValue;
     double connectivityMeasure;
     double overallDeviation;
-    double segments;
+    int segments;
 
     public Chromosome(PixelDirection[] pixelDirections) {
         this.pixelDirections = pixelDirections;
